rename and make command groups

+++ src/main/java/frc/robot/commands/Intake/RunIntakeWhenShooterAtRPMCmd.java
@@ -19,7 +19,7 @@ public class RunIntakeWhenShooterAtRPMCmd extends Command {
   long startTime;
   boolean setCandleBlink = true;
 
-  public RunIntakeWhenShooterAtRPMCmd(int RPM) {
+  public RunIntakeWhenShooterAtRPMCmd(double RPM) {
     // Use addRequirements() here to declare subsystem dependencies.
     shooter = Shooter.getInstance();
     intake = Intake.getInstance();
